#88 generic unique validator.
